requirements.txt
Flask
flake8
pip
pylint
pytest
pytest-flask
mysql-connector-python